Distributacalcul: Probability Distribution Functions

The Distributacalcul provides three general types of functions :

Functions related to continuous, discrete, and compound probability distributions.

Functions related to bivariate copulas.

Shiny applications for visualising these distributions and copulas.

Probability Distribution Functions

These include functions for calculating various moments, risk measures, and distribution functions.

Copula Functions

These include functions for calculating distribution functions and simulating observations from the copulas.

See also

Author

Maintainer: Alec James van Rassel alec.van-rassel.1@ulaval.ca [copyright holder]

Authors:

Gabriel Crépeault-Cauchon [conceptor]

Other contributors:

Étienne Marceau [teacher, scientific advisor]

Hélène Cossette [teacher, scientific advisor]

Laboratoire Act & Risk [funder, supporting host]

École d'actuariat de l'Université Laval [funder, host institution, university place]

Natural Sciences and Engineering Research Council of Canada [funder]

Marc-André Devost [conceptor]