Inverse Gaussian Distribution

Inverse Gaussian distribution with mean \(\mu\) and shape parameter \(\beta\).

expValIG(mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

varIG(mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

expValLimIG(d, mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

expValTruncIG(
  d,
  mean,
  shape = dispersion * mean^2,
  dispersion = shape/mean^2,
  less.than.d = TRUE
)

stopLossIG(d, mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

meanExcessIG(d, mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

VatRIG(kap, mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

TVatRIG(kap, mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

mgfIG(t, mean, shape = dispersion * mean^2, dispersion = shape/mean^2)

Arguments

mean

mean (location) parameter \(\mu\), must be positive.

shape

shape parameter \(\beta\), must be positive

dispersion

alternative parameterization to the shape parameter, dispersion = 1 / rate.

d

cut-off value.

less.than.d

logical; if TRUE (default) truncated mean for values <= d, otherwise, for values > d.

kap

probability.

t

t.

Value

Function :

expValIG gives the expected value.

varIG gives the variance.

expValLimIG gives the limited mean.

expValTruncIG gives the truncated mean.

stopLossIG gives the stop-loss.

meanExcessIG gives the mean excess loss.

VatRIG gives the Value-at-Risk.

TVatRIG gives the Tail Value-at-Risk.

mgfIG gives the moment generating function (MGF).

Invalid parameter values will return an error detailing which parameter is problematic.

Details

The Inverse Gaussian distribution with

Note

Function VatRIG is a wrapper for the qinvgauss function from the statmod package.

Examples

expValIG(mean = 2, shape = 5)
#> [1] 2

varIG(mean = 2, shape = 5)
#> [1] 10

expValLimIG(d = 2, mean = 2, shape = 5)
#> [1] 1.083508

expValTruncIG(d = 2, mean = 2, shape = 5)
#> [1] -0.3747381

stopLossIG(d = 2, mean = 2, shape = 5)
#> [1] 0.916492

meanExcessIG(d = 2, mean = 2, shape = 5)
#> [1] 3.383425

VatRIG(kap = 0.99, mean = 2, shape = 5)
#> [1] 15.7671

TVatRIG(kap = 0.99, mean = 2, shape = 5)
#> [1] 8861.182

mgfIG(t = 1, mean = 2, shape = .5)
#> [1] 54.59815